refactor: extract password reset form, add validation and email existence check
