+++ public/locales/en/common.json
@@ -113,7 +113,8 @@
       "genderInvalid": "Please select valid gender",
       "passwordsMustMatch": "Passwords must match!",
       "usernameTaken": "Username is already taken",
-      "emailTaken": "This email is already registered"
+      "emailTaken": "This email is already registered",
+      "emailNotFound": "This email is not registered"
     },
     "success": "User registered successfully.",
     "reset": {
