+++ public/locales/uk/common.json
@@ -113,7 +113,8 @@
       "genderInvalid": "Оберіть дійсну стать",
       "passwordsMustMatch": "Паролі повинні співпадати!",
       "usernameTaken": "Це імʼя вже використовується",
-      "emailTaken": "Ця електронна адреса вже зареєстрована"
+      "emailTaken": "Ця електронна адреса вже зареєстрована",
+      "emailNotFound": "Ця пошта не зареєстрована"
     },
     "success": "Користувача успішно зареєстровано.",
     "reset": {
